Automate Qase:  218, 182, and 219 (#143)
* Automate Qase 218, 216 and 182


* Use cluster.Transitioning instead of iterating through conditions


* Remove print


---------

+++ hosted/aks/p1/p1_provisioning_test.go
@@ -24,7 +24,7 @@ var _ = Describe("P1Provisioning", func() {
 		GinkgoLogr.Info(fmt.Sprintf("Using K8s version %s for cluster %s", k8sVersion, clusterName))
 	})
 	AfterEach(func() {
-		if ctx.ClusterCleanup && cluster != nil {
+		if ctx.ClusterCleanup && (cluster != nil && cluster.ID != "") {
 			err := helper.DeleteAKSHostCluster(cluster, ctx.RancherAdminClient)
 			Expect(err).To(BeNil())
 		} else {
@@ -79,5 +79,123 @@ var _ = Describe("P1Provisioning", func() {
 			testCaseID = 176
 			updateAutoScaling(cluster, ctx.RancherAdminClient)
 		})
+
+		It("recreating a cluster while it is being deleted should recreate the cluster", func() {
+			testCaseID = 219
+
+			err := helper.DeleteAKSHostCluster(cluster, ctx.RancherAdminClient)
+			Expect(err).To(BeNil())
+
+			// Wait until the cluster begins deletion process before recreating
+			Eventually(func() bool {
+				exists, err := helper.ClusterExistsOnAzure(clusterName, cluster.AKSConfig.ResourceGroup)
+				Expect(err).To(BeNil())
+				return exists
+			}, "1m", "5s").Should(BeFalse())
+
+			cluster, err = helper.CreateAKSHostedCluster(ctx.RancherAdminClient, clusterName, ctx.CloudCred.ID, k8sVersion, location, nil)
+			Expect(err).To(BeNil())
+
+			// wait until the error is visible on the provisioned cluster
+			Eventually(func() bool {
+				cluster, err = ctx.RancherAdminClient.Management.Cluster.ByID(cluster.ID)
+				Expect(err).To(BeNil())
+				return cluster.State == "provisioning" && cluster.Transitioning == "error" && strings.Contains(cluster.TransitioningMessage, "an AKSClusterConfig exists with the same name")
+			}, "30s", "2s").Should(BeTrue())
+
+			cluster, err = helpers.WaitUntilClusterIsReady(cluster, ctx.RancherAdminClient)
+			Expect(err).To(BeNil())
+		})
+
 	})
+
+	It("deleting a cluster while it is in creation state should delete it from rancher and cloud console", func() {
+		testCaseID = 218
+		var err error
+		cluster, err = helper.CreateAKSHostedCluster(ctx.RancherAdminClient, clusterName, ctx.CloudCred.ID, k8sVersion, location, nil)
+		Expect(err).To(BeNil())
+
+		// Wait for the cluster to appear on cloud console before deleting it
+		Eventually(func() bool {
+			exists, err := helper.ClusterExistsOnAzure(clusterName, cluster.AKSConfig.ResourceGroup)
+			// ignore the error that occurs when resource group or cluster could not be found
+			if err != nil {
+				if strings.Contains(err.Error(), fmt.Sprintf("Resource group '%s' could not be found", cluster.AKSConfig.ResourceGroup)) || strings.Contains(err.Error(), "not found") {
+					err = nil
+				}
+			}
+			Expect(err).To(BeNil())
+			return exists
+		}, "1m", "5s").Should(BeTrue())
+
+		cluster, err = ctx.RancherAdminClient.Management.Cluster.ByID(cluster.ID)
+		Expect(err).To(BeNil())
+		err = helper.DeleteAKSHostCluster(cluster, ctx.RancherAdminClient)
+		Expect(err).To(BeNil())
+
+		// Wait until the cluster finishes provisioning and then begins deletion process
+		Eventually(func() bool {
+			exists, err := helper.ClusterExistsOnAzure(clusterName, cluster.AKSConfig.ResourceGroup)
+			if err != nil {
+				if strings.Contains(err.Error(), fmt.Sprintf("Resource group '%s' could not be found", cluster.AKSConfig.ResourceGroup)) || strings.Contains(err.Error(), "not found") {
+					err = nil
+				}
+			}
+			Expect(err).To(BeNil())
+			return exists
+		}, "10m", "10s").Should(BeFalse())
+
+		// Keep the cluster variable as is so that there is no error in AfterEach; failed delete operation will return an empty cluster
+		cluster, err = ctx.RancherAdminClient.Management.Cluster.ByID(cluster.ID)
+		Expect(err).To(HaveOccurred())
+		Expect(err.Error()).To(ContainSubstring("not found"))
+	})
+
+	It("should not be able to select NP K8s version; CP K8s version should take precedence", func() {
+		testCaseID = 182
+
+		k8sVersions, err := helper.ListSingleVariantAKSAvailableVersions(ctx.RancherAdminClient, ctx.CloudCred.ID, location)
+		Expect(err).To(BeNil())
+		Expect(len(k8sVersions)).To(BeNumerically(">=", 2))
+		cpK8sVersion := k8sVersions[0]
+		npK8sVersion := k8sVersions[1]
+
+		GinkgoLogr.Info(fmt.Sprintf("Using NP K8s version: %s and CP K8s version: %s", npK8sVersion, cpK8sVersion))
+
+		updateFunc := func(clusterConfig *aks.ClusterConfig) {
+			nodePools := *clusterConfig.NodePools
+			for i := range nodePools {
+				nodePools[i].OrchestratorVersion = &npK8sVersion
+			}
+			*clusterConfig.NodePools = nodePools
+		}
+
+		cluster, err = helper.CreateAKSHostedCluster(ctx.RancherAdminClient, clusterName, ctx.CloudCred.ID, cpK8sVersion, location, updateFunc)
+		Expect(err).To(BeNil())
+
+		Expect(*cluster.AKSConfig.KubernetesVersion).To(Equal(cpK8sVersion))
+		for _, np := range cluster.AKSConfig.NodePools {
+			Expect(*np.OrchestratorVersion).To(Equal(cpK8sVersion))
+		}
+
+		cluster, err = helpers.WaitUntilClusterIsReady(cluster, ctx.RancherAdminClient)
+		Expect(err).To(BeNil())
+
+		Eventually(func() bool {
+			GinkgoLogr.Info("Waiting for the k8s upgrade to appear in AKSStatus.UpstreamSpec...")
+			clusterState, err := ctx.RancherAdminClient.Management.Cluster.ByID(cluster.ID)
+			Expect(err).To(BeNil())
+			if *clusterState.AKSStatus.UpstreamSpec.KubernetesVersion != cpK8sVersion {
+				return false
+			}
+
+			for _, np := range clusterState.AKSStatus.UpstreamSpec.NodePools {
+				if *np.OrchestratorVersion != cpK8sVersion {
+					return false
+				}
+			}
+			return true
+		}, "5m", "5s").Should(BeTrue(), "Failed while waiting for k8s upgrade.")
+	})
+
 })
